chore: attachment sanitizer 归位附件域 + 清理 forwardEventMiddleware 死代码
- createAttachmentSanitizer 移至 attachments/sanitizer.ts（base64 卫生不变量属附件域知识，由 Runner 无条件装配；不可拔插 ≠ 该住 session 层）
- session/event-middlewares.ts 只留 session 层不变量（log/persist）
- forwardEventMiddleware 零引用删除（pipeline sink 直接承担转发）
