+++ docs/official/architecture.md
@@ -28,7 +28,7 @@
 - **写入互斥**：`FileWriteMutex` 由装配点创建唯一实例，经 ProjectStore（→ SkillStore/AgentStore）、ProjectManager、SessionManager 全链路注入；`write_file`、`edit_file`、`append_changelog`、`move_file`、`copy_file` 共享之，避免同一文件并发写导致内容丢失；`generate_image` 写入唯一时间戳文件名（`.spherse/generated-images/{utc}-{hex}.{ext}`），无同路径并发，不使用 mutex
 - **删除 agent**：compound operation，由 `ProjectRuntime.deleteAgent` 协调——SessionRuntime.evictAgent 清理活跃 session、mcp capability invalidate 关闭该 agent 的 MCP 连接缓存、各 capability 的 onAgentDeleted 清理（trigger capability 删除该 agent 的 trigger 状态）、ProjectManager.deleteAgent 删除数据
 - **Agent 配置热重载**：用户编辑 agent profile 后（经 `ProjectRuntime.updateAgent` → `ProjectManager.updateAgent`），`ProjectStore` 发出 `agent_updated` 事件。`SessionManager` 构造时订阅该事件，对所有属于该 agent 的活跃 `LiveSession` 调用 `markReloadPending()`。下次 `sendMessage` 时检测到 pending 标记，先执行 `applyReload()`——从磁盘重新读取 profile，重建 systemPrompt + tools + streamFn（含 timePerception 包装），并经 `turnHooks.onReload()` 通知能力重置 memo 状态（如 mcp 合并标记）——再继续正常 prompt 流程。reload 失败时保留旧配置不中断对话。进行中的响应不受影响，配置在下一轮用户消息时才生效
-- **附件处理（Attachment）**：`AttachmentProcessor` 端口定义在 `kernel/attachments.ts`，处理器由 capability 贡献（`attachmentsCapability` 贡献 image 处理器——从磁盘路径读取 base64 ImageContent；无全局注册表，新增类型 = 新 capability）。`AgentRunner.sendMessage` 接收可选 `Attachment[]`，经 `prepareAttachmentUserMessage(text, attachments, root, processors)`（processors 由 RuntimeDeps 从 capabilities 聚合）组装为多模态 user message 后调用 `agent.prompt()`。base64 仅在本轮 LLM 调用瞬间存在——capability 可另经 `eventMiddlewares` 贡献 per-session 事件流 middleware（Runner 链固定为 `log → capability middlewares → sanitizer? → persist`，日志恒在链头、持久化恒在链尾）；attachment sanitizer 是消息管线的 base64 卫生不变量（非可选能力），将引用了 full user message 的事件中的 base64 剥离为纯文本 + `_attachments` 路径引用；`finally` 块中将 agent 内存与 MessageLog 中的 full message 替换为 stripped 版本，确保轮后内存不含 base64。附件文件落盘于 `.spherse/attachments/`（经 `attachments` category 纳入 LLM/server 读白名单，但不可写——上传走专用 `POST /attachments` 路由）。compaction 同为 capability（`compactionCapability` 贡献 afterTurn 纯变换 + 落库；`initForRestore` 对压缩记录的恢复逻辑留在 session 层——属持久层格式兼容而非能力执行，未注册该能力时旧会话 digest 仍可恢复、新 turn 不再压缩）
+- **附件处理（Attachment）**：`AttachmentProcessor` 端口定义在 `kernel/attachments.ts`，处理器由 capability 贡献（`attachmentsCapability` 贡献 image 处理器——从磁盘路径读取 base64 ImageContent；无全局注册表，新增类型 = 新 capability）。`AgentRunner.sendMessage` 接收可选 `Attachment[]`，经 `prepareAttachmentUserMessage(text, attachments, root, processors)`（processors 由 RuntimeDeps 从 capabilities 聚合）组装为多模态 user message 后调用 `agent.prompt()`。base64 仅在本轮 LLM 调用瞬间存在——capability 可另经 `eventMiddlewares` 贡献 per-session 事件流 middleware（Runner 链固定为 `log → capability middlewares → sanitizer? → persist`，日志恒在链头、持久化恒在链尾；log/persist 是 session 层不变量，住在 `session/event-middlewares.ts`）；attachment sanitizer 是附件域的 base64 卫生不变量（非可选能力、不可拔插，但归属附件域——`attachments/sanitizer.ts`，由 Runner 无条件装配），将引用了 full user message 的事件中的 base64 剥离为纯文本 + `_attachments` 路径引用；`finally` 块中将 agent 内存与 MessageLog 中的 full message 替换为 stripped 版本，确保轮后内存不含 base64。附件文件落盘于 `.spherse/attachments/`（经 `attachments` category 纳入 LLM/server 读白名单，但不可写——上传走专用 `POST /attachments` 路由）。compaction 同为 capability（`compactionCapability` 贡献 afterTurn 纯变换 + 落库；`initForRestore` 对压缩记录的恢复逻辑留在 session 层——属持久层格式兼容而非能力执行，未注册该能力时旧会话 digest 仍可恢复、新 turn 不再压缩）
 - **Skill 系统**：skill 分三层——builtin skill（来自 `@spherse/presets` 的 `PRESET_SKILL_SOURCES`，内存只读，随 app 升级更新）、project skill（`.spherse/skills/*/SKILL.md`，磁盘读写，用户自建）和 agent-level skill（`.spherse/agents/{slug}/skills/*/SKILL.md`，per-agent 私有 skill，由 `AgentStore` 的 lazy `SkillStore` 管理，无 builtin 来源）。三者均为 YAML frontmatter + Markdown body。skill capability 在 `contextBlocks` 中合并三层 skill catalog，按 name 去重，优先级 agent-level > project > builtin；`load_skill` 工具按同一优先级解析。`SkillDefinition` 的 `source` 字段（`builtin` 或 `project`）标识来源。project/agent-level skill 目录可携带附加文件（references/*、scripts/* 等），`SkillStore` 递归枚举填入 `SkillDefinition.files`，`load_skill` 加载时输出文件清单供 agent 用 `read_file` 读取。`SkillStore` 同时拥有 skill 的写逻辑（createSkill/installSkill，经 FileWriteMutex 串行化，zip-slip 防护，冲突报 `ConflictError` → server 409）
 - **触发器系统**：`TriggerManager` 管理 trigger 配置读取与触发执行（依赖 kernel `SessionPort`，经装配点的懒适配器与会话运行时解耦），`TimerService` 负责定时轮询（单个 `setTimeout` 链式，间隔 10 分钟，首 tick 对齐到最近 10 分钟整数倍）。`TriggerManager` 以磁盘为唯一真相源，每次 `onTimeTick()` / `onUserEvent()` 从磁盘读取各 agent 的 `triggers/index.yml`。Trigger 分 `time`（cron）和 `event`（用户事件）两种。event 有两个发送端：客户端经 ws-bus 调用 `onUserEvent`（含 UI SDK action），以及 agent 经 LLM tool `emit_trigger_event` in-core 直调。`payload` 为字符串，经 `{{payload}}` 注入；`sp:` 前缀内部保留
 - **memory 能力（per-agent）**：memory capability 为每个 agent 提供私有 JSONL 记忆存储（`.spherse/agents/{dir}/memory.jsonl`，惰性打开，append-only），贡献 `memory_save`/`memory_recall` 工具与 `<memory>` context block（注入当前 agent 最近 20 条记忆，per-agent 作用域经 `stores.forAgent` 隔离）；`MEMORY_PATH_RULE` 自带 LLM 读写裁决。agent 需在 profile `tools` 中显式启用
